refactor(model): Rename payouts module to payout_curve
I would like to move certain things that are scattered around the
`model` crate into the `payout_curve` module, so that we can associate
them with a particular kind of `payout_curve`.

The name `payouts` does not reflect the fact that the payout curve
variant chosen from the module has influence on other things, such as
the margin each party has to put up.
