+++ model/src/payout_curve/inverse/README.md
