+++ model/src/payout_curve/inverse/compat.rs
