+++ model/src/payout_curve/inverse/csr_tools.rs
@@ -1,6 +1,6 @@
-use crate::payouts::inverse::compat::To1DArray;
-use crate::payouts::inverse::compat::ToNAlgebraMatrix;
-use crate::payouts::inverse::Error;
+use crate::payout_curve::inverse::compat::To1DArray;
+use crate::payout_curve::inverse::compat::ToNAlgebraMatrix;
+use crate::payout_curve::inverse::Error;
 use itertools::Itertools;
 use ndarray::prelude::*;
 use std::ops::Mul;
