+++ model/src/payout_curve/inverse/curve.rs
@@ -1,7 +1,7 @@
-use crate::payouts::inverse::basis::BSplineBasis;
-use crate::payouts::inverse::splineobject::SplineObject;
-use crate::payouts::inverse::utils::*;
-use crate::payouts::inverse::Error;
+use crate::payout_curve::inverse::basis::BSplineBasis;
+use crate::payout_curve::inverse::splineobject::SplineObject;
+use crate::payout_curve::inverse::utils::*;
+use crate::payout_curve::inverse::Error;
 use ndarray::prelude::*;
 use ndarray::s;
 use std::cmp::max;
