+++ model/src/payout_curve/inverse/utils.rs
@@ -1,4 +1,4 @@
-use crate::payouts::inverse::Error;
+use crate::payout_curve::inverse::Error;
 use ndarray::prelude::*;
 use ndarray::s;
 use std::cmp::Ordering;
