+++ model/src/payout_curve/prop_compose.rs
